autotest: run_skylab_suite: improve provision suite creation logging
This line previously logged the message "Creating X dummy tasks.", which
is misleading because this is not actually the point in which the tasks
are created. And in fact, during suite wait tasks, this logging message
appears even though no new tasks are created.

BUG=chromium:957741
TEST=None

Change-Id: I098458a95266bf7420dc8948ba5c319ae759e529
Reviewed-on: https://chromium-review.googlesource.com/1588487
Commit-Ready: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>