Add test to verify that RPCs can be called
The FAFT test framework relies on RPC communication.
This test serves to check that the RPC infrastructure works,
and that all of the expected RPC functions are exposed.

Currently we do not test the effects of the RPCs. We only test that
the RPC server exposes the expected functions with the expected
arguments.

This test should function regardless of whether using the old XMLRPC
framework or the new gRPC framework.
See go/faft-grpc for more details.

Many more RPC functions will be added to this test in subsequent
check-ins.

This check-in also creates a new test suite "faft_infra,"
and adds it to the attribute_whitelist,
although it is not yet scheduled to be run automatically.

TEST=Run test_that for all control files.
BUG=chromium:963160
Cq-Depend: chromium:1590363

Change-Id: I76d64cf8bd3e21b4796d7f80e1b526da942e69ec
Signed-off-by: Greg Edelston <gredelston@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1589838
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>