firmware_ECCbiEeprom: add test for validating CBI EEPROM
We should ensure that the EEPROM used for CBI actually protects against
i2c writes when the write protect pin is asserted. We should also ensure
that writes work when the write protect pin in not asserted.

This test can be used to qualify new EEPROM parts for AVL.

BRANCH=none
BUG=b:119065537
TEST=with EC CL:1601060 on bobba, I tested the following:
1) manually run FAFT for Sparky and see pass result.
2) manually change test to put flash pin in incorrect state and verify
that test fails in both scenario

Change-Id: I3481f5544fd7c034b38f07d36967044f9f28ab39
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1601300
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>