[autotest] new autotest for ShowLogoutButtonInTray policy.
Tests the ShowLogoutButtonInTray policy in Chrome OS.
If the policy is set to True then the user can see the Sign Out button
on the task bar. If the policy is set to False/Not Set then users won't
see the Sign out button in the task bar.

BUG=None
TEST=This
Change-Id: I2bb5c10f3aaca2a5b9e4115e195cec48cd7a452e
Reviewed-on: https://chromium-review.googlesource.com/1589509
Commit-Ready: Ruben Zakarian <rzakarian@chromium.org>
Tested-by: Ruben Zakarian <rzakarian@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>