[autotest] policy_PasswordManager
Check to see if the password manager policy works. Will set to
True/False/None. If the policy is True or not set, the password manager
should be enabled. If the policy is set to True or False, the button
should be 'restricted' (ie not selectable) and have the managed by
policy icon next to it.

TEST=This
BUG=None

Change-Id: I5aab87923d031c6e5d1837f6793b32c265c2980b
Reviewed-on: https://chromium-review.googlesource.com/1596824
Commit-Ready: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>