autotest: change PASSED color for NA or WARN statuses
This changes the color of PASSED if the test failed with TestNAError or
TestWarn, so these failures are noticeable even if they aren't counted
as actual failures.

BUG=b:110265429
BRANCH=none
TEST=Run a test that passes, a test that fails with TestNAError, and a
test that fails with TestWarn. Make sure the passed test is still green,
the test na is purple, and the test warn is yellow.

Change-Id: I60179d222c318e022a49fc551b2065a438145a99
Reviewed-on: https://chromium-review.googlesource.com/1601714
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>