platform_ServoPowerStateController: make dut_off detection EC dependent
Currently, this test asserts whether a DUT is off by pinging it and
hoping for no response. This is noisy, but more importantly, it doesn't
actually answer the question whether the DUT is off, but rather whether
it responds to pings.
This change replaces that logic by querying the powerstate on the EC
instead, and considering _anything_ that is not S0 to be off.

BUG=None
TEST=manual testing on my desk with
servo-v4 + nocturne
servo-v4 + flapjack

both trigger properly when off

Change-Id: I7966e86b313ae5f7f9abae5cebf5a74069f504af
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1601727
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>