crash: Fix flakiness with rate files in logging_CrashSender
There was a change made recently where we added a 'state' directory to
the /var/lib/crash_sender path. In the autotest when it does the rate
test, it just picks any file in that dir to set back 25 hours; and if it
picked that directory then the test would fail. This fixes it so it will
only pick a file in that directory.

BUG=chromium:957342
TEST=Autotest passes

Change-Id: I89369934492e965a9facf7857e362b2a7529a6af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1602882
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Tim Zheng <timzheng@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>