power_WakeupRTC: Do not loop for all available RTC on the system
The test currently loops for all available rtc devices under
/sys/class/rtc/ and then checks if they are all wakeup capable or not.
There might be some rtc devices who do not want to have the wakeup
capability(like rtc-wilco-ec). In this case the test case fails
complaining that rtc-wilco-ec is not wakeup capable.
If anything wants to schedule a wakeup the default /dev/rtc device
which a symlink to the rtc0 device should be sufficient to do that.
So modify the autotest to remove the looping of all available rtc
devices on the system and only look for /dev/rtc.

BUG=b:129931495
TEST=test_that <IP> power_WakeupRTC works on sarien and eve.

Change-Id: I724144151ecbfd4eb341adb7c853c95b042a65b4
Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1614790
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Nick Crews <ncrews@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Michael Spang <spang@chromium.org>
Reviewed-by: Nick Crews <ncrews@chromium.org>