Allow multiple RPCs to share one test case config
Previously, config.py had a separate object for each RPC method,
even if that method would run the same set of tests as 50 other methods.
This was about to get unwieldy as we scaled up to 108 RPC methods.

With this change, a single test case can specify either a single
rpc_method or a list of rpc_methods.

TEST=test_that firmware_FAFTRPC.system
BUG=chromium:963160
Cq-Depend: chromium:1589838

Change-Id: I38907b1d0854d388adb6d366814cf6c054e66fc6
Signed-off-by: Greg Edelston <gredelston@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1615586
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>