Allow test cases to store and retrieve results
Test cases for firmware_FAFTRPC are defined in a dict, which specifies
arguments to be passed into RPC methods.

With this change, those arguments can be specified as the results of a
previous RPC call.

This change also adds documentation for the big RPC_CATEGORIES dict
defined in config.py, which contains definitions for all the test cases.

TEST=test_that firmware_FAFTRPC.bios
BUG=chromium:963160
Cq-Depend: chromium:1615586

Change-Id: I60c83f3b8ee4a7abbcce295bd23a81723f160a83
Signed-off-by: Greg Edelston <gredelston@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1618083
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>