servo: fix kb init issue
The keyboard handler in hdctools has a bug where it doesn't report yet
whether it's on/off in its get function, but just the raw boolean:
True/False. Therefore, set() fails, as it fails to certify that the kb
handler was successfully set to on.
This change fixes that and adds a TODO for when to revert this change.

BUG=b:131354266
TEST=manual testing,
test_that with a servo host and tot hdctools comes up fine

Change-Id: I6f1dca1bd7561ffb719da45f62363e314c1e42f4
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1620566
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>