firmware_Cr50DeviceState: check ap_is_on in trigger_s0
My octopus sometimes needs to short power button presses to get out of
s3/g3. I am not sure if this is an issue or not, but it would be good to
get some logs from the cr50 test devices on this. This changes the
device state test to press the power button a second time if the AP
doesn't turn on the first time.

Check if the AP is on after the power button press to see if it worked.
If it didn't try pressing the power button again. I'm adding this second
press, so we can get some patterns on what boards need two presses and
from what states.

BUG=b:79776008
BRANCH=none
TEST=run firmware_Cr50DeviceState on octopus

Change-Id: I61bd9a9011af5ee03b10e31c1e6f0d0ccb5af229
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1622739
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>