autotest: implement deferred monkeypatched module loading
This CL implements an extension to autotest.load(...) which is suitable
for use at module-global level, to allow for imports to be specified in
a more natural way at the top of a module without sacrificing the
deferred monkeypatch-then-import functionality.

This CL also modifies test_push.py to make use of this deferred loading,
both as an example and to test the new behavior.

BUG=None
TEST=unittest; local run of test_push.py

Change-Id: I94829806b0427dcd81aff56155422a995d2cb040
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1622576
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>