Fix syntax error in keyboard_device.py
Although this test is obsolete, these is a syntax error in
keyboard_device.py with an extra ')', introduced by chromium:415653.
This syntax error throws off code-health measurements, so it would be
good to fix.

Unused module imports are removed to make the pylinter hapy.

TEST=None
BUG=None

Change-Id: I11a208fe0a128c452af41e44a9aa71a706cb212f
Signed-off-by: Greg Edelston <gredelston@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1628958
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Chung-yih Wang <cywang@google.com>