power_WakeupRTC: Run test only on devices which are wakeup capable
The test loops through all available RTC devices on the system and then
runs the test on each of them. There might be cases where a device
is not wakeup capable. For such cases the test would fail, which it
should not.
So run the test only on devices which are wakeup capable.
Do this by looking for /sys/class/rtc/rtc<X>/device/power/wakeup file
present or not.
Also fail the test if /dev/rtc0 does not exist or does not have
wakeup capability, since there are many things in the system which
rely on /dev/rtc0 being present and have wakeup capability.

BUG=b:129931495
TEST=test_that <IP> power_WakeupRTC works on sarien and eve.

Change-Id: I01372c9eaf1f62a5ddfb1272c6109d1af3e0c355
Reviewed-on: https://chromium-review.googlesource.com/1633215
Tested-by: Azhar Shaikh <azhar.shaikh@intel.com>
Commit-Ready: Nick Crews <ncrews@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nick Crews <ncrews@chromium.org>