bluetooth: Add support to run single test in quick sanity
This change is to add support to run one of the tests in the batch by specifying that in test_name.
The idea is if we want to hook up the existing lab infra to run tests that are part of quick sanity, we could still run then separately.
The downside is that we'll lose the test time reduction that way.

BUG=b:132199238
TEST=manually run tests in different configurations.
Running command line -
test_that --autotest_dir ~/trunk/src/third_party/autotest/files/ --args
"chameleon_host=<Fizz_IP>" <DUT_IP>
bluetooth_AdapterLESanity.le_mouse_reports

Change-Id: I35f165105ebb0ea407bf6f11f5c8cc4864f90638
Signed-off-by: howardchung <howardchung@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1634651
Reviewed-by: Yoni Shavit <yshavit@chromium.org>