cheets_CTS_P: Add a way to set the number of testcases explicitly.
The data is for ignoring "not-executed" but "waived" failures.
Even if the tradefed itself does not think all tests are done,
as long as the expected number of test cases are run, we override
the "done" state.

I intentionally did not change the generator. The number of test cases
depends on the CTS version and also on the command line arguments.
We cannot and should not blindly inherit the current magic numbers.
Each control file regeneration needs careful check on these fields.

BUG=b:133913511
TEST=cheets_CTS_P.9.0_r8.x86.CtsTelephony

Change-Id: Id94c1d8f2fa0251b725ba6102b08b9fcc3fb53aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1635170
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>