Add RPC tests for updater RPCs that don't flash
We're avoiding testing the RPCs that write to flash memory to avoid
wearing out the flash. Shortly we will introduce a CL to put the updater
into "emulate mode" so that we can test the RPCs without actually
writing to flash. Then we will un-comment those RPCs in config.py.

TEST=test_that firmware_FAFTRPC.updater
BUG=chromium:963160

Change-Id: Ia45ef107a843dde48d9fceeed9160e1aa2d72b81
Signed-off-by: Greg Edelston <gredelston@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1641536
Reviewed-by: Dana Goyette <dgoyette@chromium.org>