autotest: Ignore repeated job assignment to a shard by master
shard_client so far assumed that a job can be assigned to the shard
exactly once. This was enforced by reporting the IDs of all known jobs
to the master and filtering out jobs matching those IDs in the jobs
assignment in the master.

Shards will no longer guarantee that all known jobs are reported to the
master in the heartbeat. Thus, some already assigned jobs may be
included again in the heartbeat response from the master. This CL
updates the heartbeat response handling to ignore such repeated job
assignments.

BUG=chromium:996872
TEST=unittests

Change-Id: If2c2c5171fd8fe47eff2361cd2ea8e24ec5fc7ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1643383
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>