network_WiFi_ChannelHop: Remove offensive SSID name
This makes for very awkward reading of test failures logs on stainless/
so change it to use the ssid naming convention other tests are using.
The use of build_unique_ssid() elsewhere suggests there was a problem
with recycling the same SSID across different wificells anyways (perhaps
due to leakage between cells making one test's SSID visible to another's
DUT and causing the DUT to get confused), so even if the name was
kosher, this change would still be good to have.

BUG=None
TEST=Run this test on nocturne, still passes

Change-Id: Ice6379c1a7e4f8d4aea981b37926eb9e3c548eb2
Signed-off-by: Kirtika Ruchandani <kirtika@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1651000
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Commit-Queue: Kirtika Ruchandani <kirtika@chromium.org>