firmware_Cr50OpenWhileAPOff: only press pwrbutton if ap isn't on
Only press the power button if the AP isn't already on to turn on the
device. The device should boot after cold_reset:on cold_reset:off. Check
the AP state and only press the power button if the AP doesn't turn on
automatically. Pressing the power button when the AP is on could turn
off the device.

BUG=none
BRANCH=none
TEST=run firmware_Cr50OpenWhileAPOff on octopus

Change-Id: I2325a5126382fe2d3a0dc324c096b68cbc576957
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1652044
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>