Test whether gRPC Python packages are present
Previously, we had tried to import gRPC for FAFT tests,
but we found that the modules were not installed on the autotest drone.
This caused http://b/134580573

We now have a CL to address this via SSP (crrev.com/c/1653740),
but just to ensure that we don't break production again,
this CL will confirm whether we have access to the required modules,
and log the results to stainless.
Once we have verified that the packages are available, we can remove
this code.

TEST=test_that firmware_FAFTSetup; and also try running with invalid
packages by changing lines 76 and 82.
BUG=chromium:971467

Change-Id: Ic0458efbfbefac4864b17e8a245dfd19231d75e3
Signed-off-by: Greg Edelston <gredelston@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1656151
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>