KernelErrorPaths: Add 3.18 variant of BUG message
On 3.18, the BUG message looks like this:

<4>[  231.188577] BUG: failure at ../../../../../tmp/portage/sys-kernel/chromeos-kernel-3_18-3.18-r2317/work/chromeos-kernel-3_18-3.18/drivers/misc/lkdtm.c:373/lkdtm_do_action()!

This does not fit our check, which is for

Kernel BUG at ....

The code already supports using tuples and looking for
a single pattern from the tuple.  We simply add the 3.18
version to the tuple.

BUG=chromium:966593
TEST=Ran KernelErrorPaths.BUG on Hana.
Change-Id: I3bb57e65d9b8c344552201a1085b5076644c5bb3
Reviewed-on: https://chromium-review.googlesource.com/1656518
Tested-by: Salman Qazi <sqazi@google.com>
Commit-Ready: Salman Qazi <sqazi@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>