cr50_test: use set_nocheck while recovering dut
Depending on the AP state and board, warm_reset may not be able to be
turned off. Use set_nocheck instead of set, so the test wont care if it
can't turn of warm_reset. It's possible it wont be able to be turned off
until the power button is pressed.

BUG=none
BRANCH=none
TEST=run faft_cr50 on octopus

Change-Id: I11830281a4dd8721bed3c9b58f5e035310d22cee
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1656950
Reviewed-by: Dana Goyette <dgoyette@chromium.org>