Files: Prevent extra logging function from throwing exception.
The desktopui_FilesApp test sometimes fails when it loses connection to
the autotest_ext. It will have the error like:
Unhandled KeyError: Cannot find a context with id <ID>.

This flake happens to any test that logs into Chrome. When this happens
in the extra logging collected by this test is hides the actual
exception that was thrown.

BUG=None
TEST=locally

Change-Id: Iba06409eec09a7d902429248a0ef4615a8e5ec78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1657608
Reviewed-by: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>