cr50_test: don't raise an error in after_run_once
Don't raise an error if the dut isn't pingable. Cleanup may pass and it
may not be an issue. _try_to_bring_dut_up is just supposed to quickly
get the device up so cleanup can start quickly. If the test got to this
point, the main test passed. Stop raising an error here, because it's
adds some unnecessary failures to test runs.

BUG=none
BRANCH=none
TEST=run firmware_Cr50OpenWhileAPOff on coral

Change-Id: I57f6817efc47c8971910de7bf255e97ff4e9976b
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1662752
Reviewed-by: Dana Goyette <dgoyette@chromium.org>