cheets_{CTS,GTS}: Fail if the test count was not a multiple of abi count.
CTS/GTS runs against multiple ABIs that the device supports.
For instance if a device supports x86 and x86_64, then it runs
each test case twice, once as x86 ABI and another as x86_64.

Thus, the total test count should become a multiple of the ABI count.

This is a hack to detect b/133420179 that runs only 151 test cases
when 151*2 = 302 is required, abusing the fact that 151 is an odd number.

BUG=b:133420179
TEST=cheets_CTS_P.9.0_r8.x86.CtsUsbTests or several devices

Change-Id: I5a3f0c539b087882db703e3100dcc46e6f4c5067
Reviewed-on: https://chromium-review.googlesource.com/1664197
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Ready: Kazuhiro Inaba <kinaba@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>