power_LoadTest: Fixed logging of urls for the page times
This change adds a delimiter between the unique salt and the url.
The old method used no delimiter and searched for 'http' in urls to
separate the components. This fails as many of the urls do not contain
that component. In that event, the substring operation takes only the
last character in the URL path.

BUG=b:135628022
TEST=Manual testing of the power_LoadTest, verified that the DEBUG logs
contain the URL.

Change-Id: I49ce9c9011f271a743648652b07dfd7eb665bf9b
Reviewed-on: https://chromium-review.googlesource.com/1668390
Tested-by: Brian Nemec <bnemec@chromium.org>
Commit-Ready: Brian Nemec <bnemec@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>