autotest: Fix boot detection in logging and return result
There was a bug introduced before where we no longer properly found the
boot line when searching through log files. This was because the regex
got changed to be a partial line match and the regex function was using
'match' instead of 'search'. This has also been updated to return
whether or not the boot line was found so tests can know if other boot
time logging should have an expectation of being found or not. The regex
was also updated to work with the journal log format as well as
messages.

BUG=chromium:817946,chromium:940608
TEST=autotest logging_UserCrash finds boot line now

Change-Id: Ica32c47e091d642559af641bea157e906caec55e
Reviewed-on: https://chromium-review.googlesource.com/1672613
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>