autotest: Fix flake in logging_UserCrash with startup detection
I believe that this test has been flakey because the logs got spammed to
the point the boot message was either gone, or something else was wrong
in the system so we couldn't get the logs. This updates the test so that
it recognizes if we can't find the boot entry in the logs, and then
fails with a different error in that case so we can prove out that
theory.

BUG=chromium:940608
TEST=autotest logging_UserCrash passes

Change-Id: Iee13eec08b9882dc607b1fc26d9bd1485ea37470
Reviewed-on: https://chromium-review.googlesource.com/1672614
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>