[autotest] Fix bug in lxc.install_packages
Ubuntu's version of pip is very old and causes segmentation faults
if installed.

Change the install_packages code to use the version of pip that is
recommended, this should come in the base container but if not it
should be installed by this code.

The only user I can find of this code is jetstream, I have created
an artificial test case in moblab to test that usage

TEST=local moblab testing
BUG=chromium:974384

Change-Id: Iceddb9f3153b1d098d6dab84d6420ce4c0cb183d
Reviewed-on: https://chromium-review.googlesource.com/1676410
Tested-by: Keith Haddow <haddowk@chromium.org>
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>