autotest: Sleep after setting consent state in crash_test
In this test we write out the policy/consent information to indicate
whether metrics are enabled/disabled. However, in the metrics library it
has a 1 second cache on that state. We were not sleeping after writing
this out, so it's very possible that we would get the wrong consent
state back from the metrics library. Adding a 2 second sleep after every
time we change it should resolve that problem.

BUG=chromium:978041
TEST=autotest logging_UserCrash

Change-Id: Ia59033aca411936ca998fb416dca9ed31c588731
Reviewed-on: https://chromium-review.googlesource.com/1676548
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>