firmware_Cr50DeviceState: track the time to run commands
If any of the commands timeout, the timeout and retry may take an extra
3 seconds. There are quite a few commands run to gather information when
when getting the irq counts. Track the time to run commands. Subtract it
from the cr50 time to determine the the amount of time cr50 was actually
idle. Use that to determine the minimum number of PMU interrupts.

BUG=b:79776008
BRANCH=none
TEST=run on octopus

Change-Id: I467f4712e44d77a216be61e01e0db8592f48de62
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1679038
Reviewed-by: Dana Goyette <dgoyette@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>