network_WiFi_APSupportedRates: improve DUT-transmission filters
{Broad,multi}cast frames may get re-transmitted back at the DUT, using
its own SA (but not TA). These are not under the DUT's control and
should not penalize the DUT, if they happen to be at a lower rate.

Similarly, some Ack or BlockAck frames are getting transmitted at lower
rates, but that only appears to be because the frame to which the DUT is
replying (i.e., the AP's frames) are transmitted at those rates. Also
avoid penalizing the DUT for simply Ack'ing at the same rate.

We probably could fix these sorts of problems within the AP, but that's
less important to us than getting the DUT behavior correct.

BUG=none
TEST=run it

Change-Id: I27d1cbf8ebdf13a0e55c29cead6b42df1c7bacb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1679995
Tested-by: Brian Norris <briannorris@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Neeraj Poojary <npoojary@chromium.org>