cr50_test: add support for getting cr50 TOT image
Some tests need to download the release image. Right now if it's a TOT
test, then the release image won't be with the prod signed images. It'll
be in the debug folder. In the future the TOT image will probably be
gotten from the reef builder.

BUG=b:133261091
BRANCH=none
TEST=run firmware_Cr50RejectUpdate

Change-Id: Ic6648298cec0dda2922d3e9094bc37e316401e45
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1682260
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>