FAFT: Ignore checking USB port connection when using CCD
When using the CCD setup, the DUT network is through the Type-C cable
to the servo v4 Ethernet chip. One test item in firmware_ECUsbPorts to
check the USB connection of all Type-A ports. As the Ethernet is no
longer connected to the Type-A port, there is no way to test it.
Ignore this check.

BUG=b:135691580
TEST=Ran the firwmare_ECUsbPorts on Octopus over CCD (the test
item ignored).

Change-Id: Iffd054b223b01118dda3889ef96d66fe7183a8d4
Reviewed-on: https://chromium-review.googlesource.com/1687362
Tested-by: Wai-Hong Tam <waihong@google.com>
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>