autotest: add static method wrapper for test_name_in_list predicate
This CL adds a new static method to the Suite class, even though static
methods on the Suite class are deprecated. Rationale: dynamic_suite
needs to call this method; it currently imports the Suite class from the
suite module, but does not import the entire suite module. Instead of
adding a static method I could instead import the entire suite module,
but that would require a lot more invasive changes and potential risk.

BUG=chromium:971377
TEST=unittests

Change-Id: I20e8209d558e7e871202927e8ff64a5c1f3b6b08
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1690644
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>