cheets_{CTS,GTS}: Always reinstall client autotest after failed login.
Previous workaround was to reinstall autotest when the script returned success
but the login was not happening. However, the breakge scenario is not limited
to that type. For safety we can always install a fresh autotest client when
retrying the login attempt.

BUG=b:129382439
TEST=cheets_CTS_P.9.0_r8.arm.CtsUsbTests
TEST=`rm -rf /usr/local/autotest/` and then cheets_CTS_P.9.0_r8.arm.CtsUsbTests
TEST=`cp /dev/null /usr/local/autotest/bin/autologin.py` then CtsUsbTests
TEST=`cp /dev/null /usr/local/autotest/setup_modules.py` then CtsUsbTests

Change-Id: I42b74cba4abb2c8ce9ec928971c21dbde579f72e
Reviewed-on: https://chromium-review.googlesource.com/1694863
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Ready: Kazuhiro Inaba <kinaba@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>