faft: add functions to fix fwids from modify_fwids
It's possible for previous runs of this test to have left the device
with a modified FWID (ending in ~RO or ~RW).  I plan to change the
tests to use the current fw image instead of flashing, but first it
needs to be able to fix any leftovers.

To make the code consistent, this also moves the real code of
modify_fwid into the flashrom handler, instead of the updater.

TEST=Run firmware_FWupdate and firmware_UpdateModes
BUG=chromium:967917

Change-Id: I5f50a3747e211e05147f72f6a304a4e7235b956a
Reviewed-on: https://chromium-review.googlesource.com/1699362
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Ready: Dana Goyette <dgoyette@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>