faft: Make firmware_FWupdate use current firmware
The test was previously using the images contained in the built-in
/usr/sbin/chromeos-firmwareupdate, but that could trigger an unwanted
downgrade, if the device was running a newer firmware before the test
started.

The test now undoes any modifications left over from previous runs,
at the start of the test.

There's logic to restore the firmware at the end of the test, but it
is possible for that logic to not run.  That's part of why it's
important to keep the running version unchanged.

TEST=Run firmware_FWupdate without args after applying a newer firmware
BUG=chromium:967917
Change-Id: Id412a7477926ce07a142f6061d717e19220226c0
Reviewed-on: https://chromium-review.googlesource.com/1699363
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Ready: Dana Goyette <dgoyette@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>