autotest: Fix startup detection in logging_UserCrash
This test has been flakey for a long time, and a recent change was made
to determine better the root cause and it's because we are unable to
find the boot line in the logs; likely because they have been flooded
with other problems and the message no longer exists. The purpose of the
check was to ensure crash_reporter actually ran its init and set the
core_pattern. Just checking core_pattern isn't reliable enough since it
gets tweaked in other tests. This adds a new flag file written out by
crash_reporter which we then check the timestamp to verify it was
written after boot.

BUG=chromium:940608
TEST=test_that logging_CrashSender passes (fails if flag is missing/old)

Cq-Depend: 1699434
Change-Id: I89fa5a5efe3c484e6885294ceb0422338ca28e5a
Reviewed-on: https://chromium-review.googlesource.com/1700578
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Ready: J Kardatzke <jkardatzke@chromium.org>
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Tim Zheng <timzheng@chromium.org>