power_LoadTest: Initial cleanup before adding video stats.
Cleanup of the data posted from the dut to the host system. Resolved
a todo to eliminate some of the shared code involved in the post messages.
Elimination of unneeded variables and fields unpacked via string
manipulation. Added a natural sort function so data is logged in the order
that they was added so entries like logs are ordered. Added a generic key
value handler which will be used to report the video stats in following CL's.

BUG=chromium:856267
TEST=Verified the power_LoadTest works as expected with various different
control files including the entries from the DUT.

Change-Id: I82712a9d2c23cdd5295b5b54a9f43b4a64c5e840
Signed-off-by: Brian Nemec <bnemec@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1707668
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>