autotest: Skip graphics_HwOverlays if no internal displays
Previously, `graphics_HwOverlays` failed on devices without internal
displays (e.g. chromeboxes). This CL updates the test to skip the tests
when there are no external displays.

Ideally, we should be able to run these tests if there is an external
display attached to the chromebox, but this is just a quick bandaid to
prevent the tests from failing.

BUG=chromium:984684
TEST=Verify that the tests are skipped on a chromebox and then verify
that the tests are still run on non-chromebox devices

Change-Id: I010702fad0438abc5ce47cf04fc87fd79dd51eda
Reviewed-on: https://chromium-review.googlesource.com/1709355
Tested-by: Brian Ho <hob@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>