Import setup_modules before importing grpc
Right now, firmware_FAFTSetup attempts to import grpc and grpc_tools.
Right now, both fail on the autotest lab drone.
This is weird, since both packages exist in site-packages/.

This change attempts to solve the problem by importing common, which in
turn adds site-packages/ to PATH.
It also logs the error message if the import fails, so that we can be
sure the package is not found, as opposed to, say, a version mismatch.

TEST=test_that firmware_FAFTSetup
BUG=chromium:971467

Change-Id: I49e3a806f2be35af4ec3dc7d1a5ea417773fda64
Signed-off-by: Greg Edelston <gredelston@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1719425
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>