cheet_CTS_P(moblab): Split dEQP-VK.subgroups into smaller pieces.
Some leak looks to be killing the test. Since the sharded dEQP test cannot
correctly run retries after crashed run, the previous setup could not recover
from the failure.

This CL is an attempt to circumvent the case by splitting the submoudle
further into smaller pieces, so that the instability due to long runs
can be avoided, hopefully.

BUG=b:138622686
TEST=cheets_CTS_P.arm.CtsDeqpTestCases.dEQP-VK.subgroup.s
TEST=cheets_CTS_P.arm.CtsDeqpTestCases.dEQP-VK.subgroup.quad

Change-Id: I8e55d83d38b909072a00809d30cafc1b047540f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1729170
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>