firmware_IntegratedU2F: Run test while logged in.
The new U2F implementation reads files from the current
user's home dir, so testing it requires a user to be
logged in. A follow up change will actually start
testing thw new implementation.

This change replaces the login_LoginSuccess test (which logs
in and then out immediately) with autologin.py (which does not).
This allows U2FTest to be run while logged in. The test logs
out manually at the end of the run by restarting the UI.

Internally, the autologin.py script logs in using the same
method as the login_LoginSuccess test.

This change also removes some calls to clear TPM, as they
are not necessary to run the test.

BUG=b:138621700
TEST=test_that firmware_IntegratedU2F

Change-Id: I882121e3ccfbb1c83ec36e9dd34d64c6a308788b
Reviewed-on: https://chromium-review.googlesource.com/1728750
Tested-by: Louis Collard <louiscollard@chromium.org>
Commit-Ready: Louis Collard <louiscollard@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>