faft_cr50: add test to check if TPM is manufactured
Add a test to check if the TPM is manufactured, so this will be super
obvious. TPM not being manufactured could lead to a lot of random
issues. We need to replace the board and identify what caused the
corrupted manufacturing space.

BUG=none
BRANCH=none
TEST=none

Change-Id: I5ac66702ae637990531cb174c4d78bc896a3aa53
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1731442
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>